Python bindings for ImageBufAlgo
The vast majority of ImageBufAlgo functions are hereby exposed to Python.

A small number of functions have temporarily been omitted.  These are
all rarely used, and had call signatures where I was unsure of the
correct Pythonic way to organize them.  We'll come back to this later,
but in the mean time, we have now working versions of almost everything
else in ImageBufAlgo that you can use from Python.