Build libOpenImageIO_Util containing just the utilities from src/libutil.
This is useful for apps that like the stuff in libutil but don't actually
need to read and write images or need any of the Image* classes.

Make/CMake option BUILD_OIIOUTIL_ONLY=1 will cause only this to be built,
and will not build libOpenImageIO at all (can be faster if you truly only
need to build the utils to link against an app).

Moved all the unit tests specific to libutil (and unit tests that are
related to header-only utility classes) there, instead of libOpenImageIO.