Improved removal of SHA-1 hash upon reading in oiiotool and for IBA
This is a continuation of the recent checkin related to this.  When
reading in oiiotool, or when preparing destination IB's for IBA
functions, we had removed any actual "oiio:SHA-1" metadata.  But this
patch also ensures that any hashes embedded in the "ImageDescription"
are also removed.

Also, in IBAPrep, use flags to determine how much metadata should be
copied when initializing the dst image.