IBA resize optimization
Improve the performance of IBA::resize when using separable filters with
large support (e.g., resizing an 8k image to 512 using Lanczos improves
resize performance by over 2x). The refactor of IBA::resize is primarily
to hoist all the redundant calls to filter evaluation and normalization
out of the inner loops -- for each (x,y) in the output, the x filter
taps can be computed once (rather than once for each y tap), and the y
filter taps only need to be computed and normalized once for each
scanline in the image. Also, we change to having a single IB::iterator
over all the filter taps, rather than having one for each row; this
reduces a lot of overhead.