Consistency improvement for iinfo and oiiotool --info
iinfo and oiiotool (and for each, the whole image versus subimages)
didn't all match the order and nomenclature for their presentation of
volume, deep, and per-channel data formats. This makes them all match,
and in particular fixes the fact that oiiotool didn't say that deep
files were deep (but iinfo did).