GIF input plugin fixes and improvements.
Fix subimages generation. GIF frames are treated now as sequential
windows to be drawn on canvas rather than as independent images. Also,
respect disposal method so canvas is properly disposed or saved before
drawing of the next frame. The initial canvas is transparent and as a
result all GIFs are presented as 4-channel images.

Add gif:LoopCount attribute.

Add a couple of new tests.