maketx: Fix conflict between --checknan and --fixnan.
If both were specified, it would checknan first, erroring out before
getting the chance to fix it. Now, if both commands are specified, it
does the fix first, then the check. (Not completely superfluous -- it
is conceivable that fixnan can not fully fix it, for example if there
are extended NaN regions over which the box3 solution can't fix things.
But usually, you wouldn't want to check if you fix.)