Workaround for USE_CONCURRENT_MAP seg fault on 32 bit systems with gcc 4.2.
Very few people are on 32 bit systems, let alone using a gcc so old, so
don't sweat the lower perf of turning USE_CONCURRENT_MAP off in that
case.  It's not even worth trying to track down exactly what's wrong.

Also note that all we know is that it core dumps with gcc 4.2, and is
fine with gcc 4.6 and 4.8. It's possible that we'll need to adjust the
test to also exclude other old gcc's, but we'll cross that bridge when
we get to it.