Tweak IBAPrep - usually copy src's display window, but option not to.
Previously, it just set an uninitialized dst's display window to the data
window, but that's often not right for cropped images. Usually, in fact,
it should copy the source image's display window. But for a few ops, that's
the wrong thing to do, so we give a flag option not to do it.