Bug fix: oiiotool color conversion botched inputs with MIP levels.
It should have been looping over the number of subimages and mip levels
in the destination image, not the source image, because if the -a flag
is not used, the dst image will be constructed with fewer mip levels
(i.e. 1) than the source. Looping over the source's count will run off
the end of the list.