Add maketx option to sharpen as we create MIP levels.
The idea is that production sometimes thinks the filtered textures look
a little too blurry and washed out (even though we know that they're
filtered almost exactly right). So with this, we optionally do a
sharpening step (via unsharp_mask) as we create MIP levels. This gooses
the high-frequency detail just a bit.  Example command line:

  maketx input.exr -sharpen 0.5 -filter unsharp-median -o texture.exr

Play with the sharpening amount (I like somewhere between 0.25 and 0.5)
and with whether you use the unsharp-median or (by leaving out the
-filter or setting it to an ordinary name like "lanczos3") whether you
use the usual Gaussian-based unsharp mask

This is all experimental! Use at your own risk.