SIMD instruction speedups for texture lookups
This is a big overhaul of the internals of TextureSystem to make as much
as possible use SSE (and in future updates, other SIMD instruction
sets).  We're seeing 30-60% reduction in time of the texture lookups
themselves, depending on the SSE compatibility level, number of channels
looked up, whether texture derivatives are requested, etc.  Call it
close to a 2x speedup in general. But of course, how much this will
translate into speedup of an app using TextureSystem will depend a lot
on the original percentage of the runtime of the app that was spent
doing the texture math.

This is a first stab, there is more to return to in the future. For
example, we have only SSE-optimized texture(), but really not
texture3d() or environment(). There's probably even more we can squeeze
out of texture() if we try.