OIIO_RESTRICT macro for restricted pointers.
Use on function declarations like this:

    void add1 (float* OIIO_RESTRICT A, const float* OIIO_RESTRICT B)
    {
        A[0] = B[0] + 1.0f;
        A[1] = B[1] + 1.0f;
    }

This means "A and B will never point to the same memory in a way that
could create a data dependency". Without this, it could be that B[1] is
the same memory as A[0], which means that the first line MUST execute
prior to the second in order to be correct, whereas if it knew A and B
would point to distinct memory (or at least would never interfere), the
compiler could be more creative with optimizations because those lines
could be rearranged or run in parallel.

Use with extreme caution!