testtex enhancements: --nchannels, --derivs
--nchannels lets you force the number of channels to look up (it used to
automatically use the number of channels in the map, and this is still
the default). The purpose is to be able to test when the number of channels
in the lookup differs from the number in the map.

--derivs makes the variety of texture call that looks up derivatives of the
texture as well, and outputs them as two additional images (one for ds,
one for dt). This lets us test that the derivatives are calculated correctly.