Address SSE3 vs SSSE3.
Brecht points out that we set OIIO_SIMD_SSE to 3 when detecting __SSE3__,
but actually, some instructions may require SSSE3. Oy.

My theory is that there are only a few old (only 32 bit?) architectures
that are SSE3 but *not* SSSE3, and that it's no big deal for those oddballs
to get SSE2. So my proposed solution is to set OIIO_SIMD_SSE to 3 when
we see __SSSE3__.