Python 3 support from Lukas Schrangl
Add a new USE_PYTHON3 cmake switch (and similar in the makefile
wrapper), which is off by default. By setting it to ON (e. g. run cmake
-DUSE_PYTHON3=ON .. or make USE_PYTHON3=1), the python3 bindings for
OIIO are built (independently from the python2 bindings, which are still
controlled via the USE_PYTHON switch.)

Fix py_imagespec.cpp for python3: Replace PyString_* by PyUnicode_* and
PyInt_* by PyLong_* functions, all in #if PY_MAJOR_VERSION > 3
blocks so that it is backwards compatible with python2.