Move grid.tx and checker.tx from oiio-images to testsuite/common/textures.
This is really being done because the grid.tx in oiio-images is really
old and I want to re-txmake it for better filtering, higher quality
resize, and more modern metadata. OK, but if I do that, it'll make old
branch testsuites fail horribly. So the solution is to move the texture
to the oiio tree itself. This is an icky one-time thing. Sorry.