Tweak ffmpeg reader to only open files ending in certain extensions.
This TEMPORARILY fixes the problem of this reader willing to open too
many file types best handled by other plugins.

This should eventually be replaced by a more robust solution, but as
a temporary measure this is enough to allow us to enable ffmped building
by default for the next release.