TIFF: allow writing of half (fp16) pixels. (currently disabled!!!)
This is allowed according to a 2005 standard extension by Adobe.  OIIO
has read these files for a while, but never written them (instead
silently promoting to float). By using this, we are taking a slight
chance that non-OIIO TIFF-reading software may not understand the files.
I suppose if this starts cropping up as a problem, we may revisit the
decision.