simd.h refinement: found faster load(...,3) and store(...,3) idioms.
It looks like we're trading simd for serial ops here, but looking at the
assembly output shows that gcc and clang (especially clang!) recognize
this pattern and turn it into very efficient code. Furthermore, I
benchmarked it very carefully, and this makes a definite speedup
(significantly so for the 3-float load). I still don't fully understand
why my trickery wasn't the fastest method, but the timings could not be
more clear, on multiple platforms.