TextureSystem: handle-based versions of more methods
We've already had ways to retrieve an opaque handle to a texture file,
based on the name, and versions of texture(), texture3d(), environment()
lookups based on the handle as well as the filename. For an app that
already has the handle and a thread-specific pointer, texture lookup is
faster than if the name->file mapping must be figured out and the
thread-specific pointer retrieved.

This patch extends this technique to get_texture_info(),
get_imagespec(), imagepec(), and get_texels(), adding a method for each
that takes a TextureHandle* (and for some of them also a Perthread*),
which are slightly faster than the versions that take filenames and have
to do the name lookups.