OpenEXR write_tile with AutoStride calculated the wrong strides.
When strides are automatically computed rather than explicitly supplied,
it should have assumed data layout of a "full tile", even if it was the
last tile in a row in an image whose width was not an integral multiple
of the tile width. Instead, it was clamping the width and height by the
size of the remaining data. That didn't match the behavior of read_tile
(which assumed full-tile-size layout, even for edge tiles), nor the
behavior of write_tile for other tiled formats such as TIFF.

Used the occasion to tighten up the imageio.h comments and OIIO
documention on this matter, spelling out more carefully what happens
with AutoStride in write_tile versus write_tiles.