oiiotool: gracefully handle impossible requests to save too many channels.
Before saving and encountering an error, check that the output file type
supports the number of channels in the image being saved. If not, don't
consider it an error, but rather just give a warning, and then do the
best we can. What we try to do is: (a) if "R", "G", and "B" (and also
"A", if the format supports it) are all in the list of channel names for
the image that is being saved, save just those channels; (b) if those
names are not found, just punt and save the first three channels (or 4,
if the output supports alpha).