maketx improvement: write to temp file, rename upon successful completion.
This improves robustness by eliminating the possibility of a
half-written texture file that lacks all the MIPmap levels (which might
be inexplicably slow to render if nobody realizes it's a partial file)
or is otherwise malformed in a way that leads to chaos. Basically, a
crash or failure no longer will write to the output filename, only to a
temporary file. The temp file is renamed/moved when the full file write
is complete without errors.