More IC/TS perthread fixes
App-owned Perthead pointers (the ones preallocated by create_perthread_info)
had no mechanism for clearing their per-thread microcaches of tile references
even after they were marked as "purge." That's because the actual purging
was previously done in the get_perthread_info function that was used
to retrieve the true thread-specific-pointers. This could lead to
inconsistent texture results for different threads, if an invalidate()
for that file had happened at just the right time.

The fix the strategy is that now get_perthread_info takes an optional
Perthread* argument, which if supplied is used, and only if NULL is
passed does it retrieve and/or allocate a TSP. Then in either case, it
does the other housekeeping (including purging, when requested) before
returning the pointer (either the supplied one passed in, or the
created/retrieved one).